CLeaning Warning, adding Path3d, correcting some unitary tests, adding @Pure

+++ core/math/src/test/java/org/arakhne/afc/math/graph/astar/AStarNotOrientedTest.java
@@ -99,7 +99,7 @@ public class AStarNotOrientedTest extends AbstractMathTestCase {
 	}
 	
 	private static double dist(double a, double b) {
-		return (double)Math.sqrt(a*a+b*b);
+		return Math.sqrt(a*a+b*b);
 	}
 	
 	/**
